Error on undefined template variables
Previously, a variable x would be lifted to `return x` if template_x is
not defined. This is unlikely to be the desired behavior, especially if
x is a function, and the resulting error message makes it hard to
identify the mistake. This changes it to instead produce an error
stating that template_x is undefined, which matches the behavior when
lifting constructors.